# atoma-proxy/docs/openapi.yml
openapi: 3.1.0
info:
  title: atoma-proxy
  description: ''
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8080
paths:
...
  /v1/models:
    get:
      tags:
      - Models
      summary: List models
      description: |-
        This endpoint mimics the OpenAI models endpoint format, returning a list of
        available models with their associated metadata and permissions. Each model
        includes standard OpenAI-compatible fields to ensure compatibility with
        existing OpenAI client libraries.

        # Arguments

        * `state` - The shared application state containing the list of available models

        # Returns

        Returns a JSON response containing:
        * An "object" field set to "list"
        * A "data" array containing model objects with the following fields:
          - id: The model identifier
          - object: Always set to "model"
          - created: Timestamp (currently hardcoded)
          - owned_by: Set to "atoma"
          - root: Same as the model id
          - parent: Set to null
          - max_model_len: Maximum context length (currently hardcoded to 2048)
          - permission: Array of permission objects describing model capabilities

        # Example Response

        ```json
        {
          "object": "list",
          "data": [
            {
              "id": "meta-llama/Llama-3.1-70B-Instruct",
              "object": "model",
              "created": 1730930595,
              "owned_by": "atoma",
              "root": "meta-llama/Llama-3.1-70B-Instruct",
              "parent": null,
              "max_model_len": 2048,
              "permission": [
                {
                  "id": "modelperm-meta-llama/Llama-3.1-70B-Instruct",
                  "object": "model_permission",
                  "created": 1730930595,
                  "allow_create_engine": false,
                  "allow_sampling": true,
                  "allow_logprobs": true,
                  "allow_search_indices": false,
                  "allow_view": true,
                  "allow_fine_tuning": false,
                  "organization": "*",
                  "group": null,
                  "is_blocking": false
                }
              ]
            }
          ]
        }
        ```
      operationId: models_handler
      responses:
        '200':
          description: List of available models
          content:
            application/json:
              schema: {}
        '500':
          description: Failed to retrieve list of available models
      security:
      - bearerAuth: []
  /node/registration:
    post:
      tags:
      - Node Public Address Registration
      summary: Register node
      description: |-
        This endpoint allows nodes to register or update their public address in the system.
        When a node comes online or changes its address, it can use this endpoint to ensure
        the system has its current address for routing requests.

        # Arguments

        * `state` - The shared application state containing the state manager sender
        * `payload` - The registration payload containing the node's ID and public address

        # Returns

        Returns `Ok(Json(Value::Null))` on successful registration, or an error status code
        if the registration fails.

        # Errors

        Returns `StatusCode::INTERNAL_SERVER_ERROR` if:
        * The state manager channel is closed
        * The registration event cannot be sent

        # Example Request Payload

        ```json
        {
            "node_small_id": 123,
            "public_address": "http://node-123.example.com:8080"
        }
        ```
      operationId: node_public_address_registration
      responses:
        '200':
          description: Node public address registered successfully
          content:
            application/json:
              schema: {}
        '500':
          description: Failed to register node public address
  /v1/chat/completions:
    post:
      tags:
      - Chat
      summary: Create chat completion
      description: |-
        This function processes chat completion requests by determining whether to use streaming
        or non-streaming response handling based on the request payload. For streaming requests,
        it configures additional options to track token usage.

        # Arguments

        * `metadata`: Extension containing request metadata (node address, ID, compute units, etc.)
        * `state`: The shared state of the application
        * `headers`: The headers of the request
        * `payload`: The JSON payload containing the chat completion request

        # Returns

        Returns a Response containing either:
        - A streaming SSE connection for real-time completions
        - A single JSON response for non-streaming completions

        # Errors

        Returns an error status code if:
        - The request processing fails
        - The streaming/non-streaming handlers encounter errors
        - The underlying inference service returns an error
      operationId: chat_completions_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
        required: true
      responses:
        '200':
          description: Chat completions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
  /v1/confidential/chat/completions:
    post:
      tags:
      - Confidential Chat
      summary: Create confidential chat completion
      description: |-
        This handler processes chat completion requests in a confidential manner, providing additional
        encryption and security measures for sensitive data processing. It supports both streaming and
        non-streaming responses while maintaining data confidentiality through AEAD encryption and TEE hardware,
        for full private AI compute.

        # Arguments

        * `metadata` - Extension containing request metadata including:
          * `endpoint` - The API endpoint being accessed
          * `node_address` - Address of the inference node
          * `node_id` - Identifier of the selected node
          * `num_compute_units` - Available compute units
          * `selected_stack_small_id` - Stack identifier
          * `salt` - Optional salt for encryption
          * `node_x25519_public_key` - Optional public key for encryption
          * `model_name` - Name of the AI model being used
        * `state` - Shared application state (ProxyState)
        * `headers` - HTTP request headers
        * `payload` - The chat completion request body

        # Returns

        Returns a `Result` containing either:
        * An HTTP response with the chat completion result
        * A streaming SSE connection for real-time completions
        * A `StatusCode` error if the request processing fails

        # Errors

        Returns `StatusCode::BAD_REQUEST` if:
        * The 'stream' field is missing or invalid in the payload

        Returns `StatusCode::INTERNAL_SERVER_ERROR` if:
        * The inference service request fails
        * Response processing encounters errors
        * State manager updates fail

        # Security Features

        * Utilizes AEAD encryption for request/response data
        * Supports TEE (Trusted Execution Environment) processing
        * Implements secure key exchange using X25519
        * Maintains confidentiality throughout the request lifecycle

        # Example

        ```rust,ignore
        let response = confidential_chat_completions_handler(
            Extension(metadata),
            State(state),
            headers,
            Json(payload)
        ).await?;
        ```
      operationId: confidential_chat_completions_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
        required: true
      responses:
        '200':
          description: Confidential chat completions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
  /v1/embeddings:
    post:
      tags:
      - Embeddings
      summary: Create embeddings
      description: |-
        This endpoint follows the OpenAI API format for generating vector embeddings from input text.
        The handler receives pre-processed metadata from middleware and forwards the request to
        the selected node.

        Note: Authentication, node selection, and initial request validation are handled by middleware
        before this handler is called.

        # Arguments
        * `metadata` - Pre-processed request metadata containing node information and compute units
        * `state` - The shared proxy state containing configuration and runtime information
        * `headers` - HTTP headers from the incoming request
        * `payload` - The JSON request body containing the model and input text

        # Returns
        * `Ok(Response)` - The embeddings response from the processing node
        * `Err(StatusCode)` - An error status code if any step fails

        # Errors
        * `INTERNAL_SERVER_ERROR` - Processing or node communication failures
      operationId: embeddings_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmbeddingRequest'
        required: true
      responses:
        '200':
          description: Embeddings generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEmbeddingResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
  /v1/confidential/embeddings:
    post:
      tags:
      - Confidential Embeddings
      summary: Create confidential embeddings
      description: |-
        This endpoint follows the OpenAI API format for generating vector embeddings from input text,
        but with confidential processing (through AEAD encryption and TEE hardware).
        The handler receives pre-processed metadata from middleware and forwards the request to
        the selected node.

        Note: Authentication, node selection, initial request validation and encryption
        are handled by middleware before this handler is called.

        # Arguments
        * `metadata` - Pre-processed request metadata containing node information and compute units
        * `state` - The shared proxy state containing configuration and runtime information
        * `headers` - HTTP headers from the incoming request
        * `payload` - The JSON request body containing the model and input text

        # Returns
        * `Ok(Response)` - The embeddings response from the processing node
        * `Err(StatusCode)` - An error status code if any step fails

        # Errors
        * `INTERNAL_SERVER_ERROR` - Processing or node communication failures
      operationId: confidential_embeddings_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmbeddingRequest'
        required: true
      responses:
        '200':
          description: Confidential embeddings generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEmbeddingResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ChatCompletionRequest:
      type: object
      required:
      - model
      - messages
      properties:
        frequency_penalty:
          type:
          - number
          - 'null'
          format: float
          description: |-
            Number between -2.0 and 2.0. Positive values penalize new tokens based on their
            existing frequency in the text so far
        function_call:
          description: Controls how the model responds to function calls
        functions:
          type:
          - array
          - 'null'
          items: {}
          description: A list of functions the model may generate JSON inputs for
        logit_bias:
          type:
          - object
          - 'null'
          description: Modify the likelihood of specified tokens appearing in the completion
          additionalProperties:
            type: number
            format: float
          propertyNames:
            type: string
        max_tokens:
          type:
          - integer
          - 'null'
          format: int32
          description: The maximum number of tokens to generate in the chat completion
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionMessage'
          description: A list of messages comprising the conversation so far
        model:
          type: string
          description: ID of the model to use
        n:
          type:
          - integer
          - 'null'
          format: int32
          description: How many chat completion choices to generate for each input message
        presence_penalty:
          type:
          - number
          - 'null'
          format: float
          description: |-
            Number between -2.0 and 2.0. Positive values penalize new tokens based on
            whether they appear in the text so far
        response_format:
          description: The format to return the response in
        seed:
          type:
          - integer
          - 'null'
          format: int64
          description: If specified, our system will make a best effort to sample deterministically
        stop:
          type:
          - array
          - 'null'
          items:
            type: string
          description: Up to 4 sequences where the API will stop generating further tokens
        stream:
          type:
          - boolean
          - 'null'
          description: Whether to stream back partial progress
        temperature:
          type:
          - number
          - 'null'
          format: float
          description: What sampling temperature to use, between 0 and 2
        tool_choice:
          description: Controls which (if any) tool the model should use
        tools:
          type:
          - array
          - 'null'
          items: {}
          description: A list of tools the model may call
        top_p:
          type:
          - number
          - 'null'
          format: float
          description: An alternative to sampling with temperature
        user:
          type:
          - string
          - 'null'
          description: A unique identifier representing your end-user
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...